Correct fix for sinq center command
numpy changed the promotion/overflow handling for fixlength int,
resulting in negative square.

Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37373
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Tested-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>

+++ nicos_sinq/sxtal/commands/__init__.py
@@ -418,7 +418,7 @@ def Max(dev, step, counter=None, maxpts=None, **preset):
         session.log.error('%s not found, cannot maximize against is', counter)
         return False
     preset['temporary'] = True
-    counts = np.zeros((maxpts,), dtype='int32')
+    counts = np.zeros((maxpts,), dtype='int')
     positions = np.zeros((maxpts,), dtype='float32')
     positions[half_point] = dev.read()
     _count(*(), **preset)
